# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Golf Court Reservation API",
        "description": "Documentación de la API para el sistema de reservas de canchas de golf.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1/api",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/reservations": {
            "get": {
                "tags": [
                    "Reservations"
                ],
                "summary": "Obtener listado de reservas",
                "description": "Retorna una lista paginada de reservas con filtros opcionales por estado, fecha, club_id y player_id.",
                "operationId": "getReservationsList",
                "parameters": [
                    {
                        "name": "perPage",
                        "in": "query",
                        "description": "Cantidad de elementos por página (máximo 100)",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 25
                        }
                    },
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Filtrar por estado de la reserva",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "date",
                        "in": "query",
                        "description": "Filtrar por fecha (YYYY-MM-DD)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "club_id",
                        "in": "query",
                        "description": "Filtrar por ID del club",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "player_id",
                        "in": "query",
                        "description": "Filtrar por ID del jugador",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Operación exitosa"
                    }
                }
            },
            "post": {
                "tags": [
                    "Reservations"
                ],
                "summary": "Crear una nueva reserva",
                "description": "Registra una nueva reserva de cancha de golf.",
                "operationId": "storeReservation",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "club_id",
                                    "player_id",
                                    "date",
                                    "start_time",
                                    "end_time",
                                    "players_count",
                                    "status"
                                ],
                                "properties": {
                                    "club_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "player_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2026-08-26"
                                    },
                                    "start_time": {
                                        "type": "string",
                                        "example": "11:00"
                                    },
                                    "end_time": {
                                        "type": "string",
                                        "example": "13:00"
                                    },
                                    "players_count": {
                                        "type": "integer",
                                        "example": 2
                                    },
                                    "status": {
                                        "type": "string",
                                        "example": "pending"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Reserva creada exitosamente"
                    },
                    "422": {
                        "description": "Error de validación (campos faltantes o con formato incorrecto)"
                    },
                    "409": {
                        "description": "Conflicto de horario: ya existe una reserva que se solapa"
                    }
                }
            }
        },
        "/reservations/{reservation}": {
            "get": {
                "tags": [
                    "Reservations"
                ],
                "summary": "Mostrar una reserva específica",
                "description": "Retorna la información detallada de una reserva mediante su ID.",
                "operationId": "getReservationById",
                "parameters": [
                    {
                        "name": "reservation",
                        "in": "path",
                        "description": "ID de la reserva",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Operación exitosa"
                    },
                    "404": {
                        "description": "Reserva no encontrada"
                    }
                }
            },
            "put": {
                "tags": [
                    "Reservations"
                ],
                "summary": "Actualizar una reserva existente",
                "description": "Modifica los datos de una reserva existente de forma parcial o total.",
                "operationId": "updateReservation",
                "parameters": [
                    {
                        "name": "reservation",
                        "in": "path",
                        "description": "ID de la reserva",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "club_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "player_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2026-08-26"
                                    },
                                    "start_time": {
                                        "type": "string",
                                        "example": "11:00"
                                    },
                                    "end_time": {
                                        "type": "string",
                                        "example": "13:00"
                                    },
                                    "players_count": {
                                        "type": "integer",
                                        "example": 2
                                    },
                                    "status": {
                                        "type": "string",
                                        "example": "confirmed"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Reserva actualizada con éxito"
                    },
                    "404": {
                        "description": "Reserva no encontrada"
                    },
                    "422": {
                        "description": "Error de validación"
                    },
                    "409": {
                        "description": "Conflicto de horario: ya existe otra reserva que se solapa"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Reservations"
                ],
                "summary": "Eliminar una reserva",
                "description": "Elimina una reserva del sistema.",
                "operationId": "deleteReservation",
                "parameters": [
                    {
                        "name": "reservation",
                        "in": "path",
                        "description": "ID de la reserva",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Reserva eliminada exitosamente (sin contenido)"
                    },
                    "404": {
                        "description": "Reserva no encontrada"
                    }
                }
            }
        }
    },
...
}